pyproject.toml
[tool.poetry]
name = "vextract"
version = "1.0.0"
description = "Code associated with CSH resistivity analysis"
authors = ["James J. Hamlin <jhamlin@ufl.edu>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
ipympl = "^0.9.2"
jupyterlab = "^3.4.7"
jupyterlab-vim = "^0.15.1"
matplotlib = "^3.6.0"
pandas = "^1.5.0"
svgpathtools = "^1.5.1"
numpy-ext = "^0.9.8"
webcolors = "^1.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
